Cloud database

+++ src/app/pizza-order/pizza-order.component.ts
@@ -4,6 +4,8 @@ import { Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { PizzaOrderFirebaseService } from '../service/pizza-order-firebase.service';
 import { Router } from '@angular/router';
+import { PizzaOrder } from '../pizza-order';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-pizza-order',
@@ -15,7 +17,7 @@ export class PizzaOrderComponent implements OnInit {
   messages = [];
 
   constructor(private messageService: MessageService, private pizzaOrderFirebaseService: PizzaOrderFirebaseService,
-    private router: Router,) {}
+    private router: Router, private toastr: ToastrService) {}
 
   ngOnInit(): void {
     // Observable Used
@@ -25,13 +27,24 @@ export class PizzaOrderComponent implements OnInit {
     //   }
     // });
 
-    this.pizzaOrderFirebaseService.getPizzaOrderList().snapshotChanges().pipe(
-      map(changes => 
-        changes.map(c => 
-          ({key: c.payload.key, ...c.payload.val()})
-          )
-        )
-    ).subscribe(orders => this.messages = orders);
+    // Firebase database
+    // this.pizzaOrderFirebaseService.getPizzaOrderList().snapshotChanges().pipe(
+    //   map(changes => 
+    //     changes.map(c => 
+    //       ({key: c.payload.key, ...c.payload.val()})
+    //       )
+    //     )
+    // ).subscribe(orders => this.messages = orders);
+
+    // Firebase Cloud database
+    this.pizzaOrderFirebaseService.getPizzaOrderList().subscribe(actionArrays => {
+      this.messages = actionArrays.map(item => {
+        return {
+          id: item.payload.doc.id,
+          ...(item.payload.doc.data() as PizzaOrder),
+        } as PizzaOrder
+      }) 
+    });
   }
 
   removeAllOrders() {
@@ -39,11 +52,13 @@ export class PizzaOrderComponent implements OnInit {
   }
 
   removeOrder(key: string) {
-    console.log(this.messages);
-    this.pizzaOrderFirebaseService.deletePizzaOrder(key);
+    if (confirm('Are you sure to delete this record?')) {
+      this.pizzaOrderFirebaseService.deletePizzaOrder(key);
+      this.toastr.warning('Deleted Successfully', 'Pizza Reg.');
+    }
   }
 
-  updateOrder(id: string) {
+  editOrder(id: string) {
     this.router.navigate(['/create', id]);
   }
